fix: fix bug where floating IPs were incorrectly getting created in the default resource group (#580)
BREAKING CHANGE: If you are upgrading to this version, and have set `enable_floating_ip` to `true`, or are passing values to the `secondary_floating_ips` list, the floating IPs are going to be deleted and recreated in the correct resource group. Please plan accordingly before upgrading incase this change will cause disruption for whatever is using the floating IPs.